+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added optional "ephemeral" parameter to specifiy whether ephemeral sessions (previous standard!) should be used or not. **NOTE**: Default behaviour was changed to `ephemeral = false`!
 * Fixed issue where Login attempts would appear as "canceled" on Firefox
+* Added Toast warning message if no browser is installed, added error code to inform about missing browser
 
 ## [0.3.0] - (28th January 2022)